% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_priors.R
\name{compare_priors}
\alias{compare_priors}
\title{Compare priors}
\usage{
compare_priors(coefficients_priortype_1, coefficients_priortype_0, Effect)
}
\arguments{
\item{coefficients_priortype_1}{dataframe with coefficients from regression output with priortype 1}

\item{coefficients_priortype_0}{dataframe with coefficients from regression output with priortype 0}

\item{Effect}{Effect from intervention_output}
}
\value{
plot
}
\description{
Compare priors
}
