% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_interventions.R
\name{make_interventions}
\alias{make_interventions}
\title{Make interventions}
\usage{
make_interventions(data, Effect, Kwartaaleffect, Runs)
}
\arguments{
\item{data}{a dataframe with colnames id, Interventie and K1}

\item{Effect}{value of the intervention effect on referrals}

\item{Kwartaaleffect}{value of the quarterly change in referrals}

\item{Runs}{number of simulations}
}
\value{
dataframe with 9 columns and runs * id * 12 rows. containing # of referrals per quarter per id
}
\description{
Make interventions
}
