% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_participants.R
\name{make_participants}
\alias{make_participants}
\title{Make participants
simulates an intervention and control group}
\usage{
make_participants(IP, CP)
}
\arguments{
\item{IP}{the number of intervention participants}

\item{CP}{the number of control participants}
}
\value{
data frame with participants, indicated with interventie (yes/no) and mean number of verwijzingen.
}
\description{
Make participants
simulates an intervention and control group
}
