% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_samples.R
\name{plot_samples}
\alias{plot_samples}
\title{Plot sample distribution}
\usage{
plot_samples(samples, Grens, Effect)
}
\arguments{
\item{samples}{dataframe with posterior samples}

\item{Grens}{boundary indicator: estimates smaller than?}

\item{Effect}{What intervention effect is plotted?}
}
\value{
density plots for a specific quarter with percentage indicator
}
\description{
Plot sample distribution
}
