% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_regressions.R
\name{run_regressions}
\alias{run_regressions}
\title{run regressions
this function runs a bayesian regression with
Formula: "Verwijzingen ~ 0 + Intercept + id + Kwartaal + DiD_9 + DiD_10 + DiD_11 + DiD_12"}
\usage{
run_regressions(data, prior_type)
}
\arguments{
\item{data}{dataframe with interventions}

\item{prior_type}{priors for intercept and DiD variables. 1 = non-informative prior, 0 = informative prior}
}
\value{
list containing fixed effect coefficients and samples
}
\description{
run regressions
this function runs a bayesian regression with
Formula: "Verwijzingen ~ 0 + Intercept + id + Kwartaal + DiD_9 + DiD_10 + DiD_11 + DiD_12"
}
