% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatebayes-package.R
\docType{package}
\name{simulatebayes-package}
\alias{simulatebayes}
\alias{simulatebayes-package}
\title{simulatebayes: This packages simulates data an experiment}
\description{
This packages simulates data and runs bayesian regression for an financing experiment among GP's. The package simulates data on GP's that refer patients to hospitals. The referrals are simulated for 12 quarters. The intervention is a new way of financing and from quarter 9 up until quarter 12.
}
\keyword{internal}
